+++ admin/src/scenes/structure/view/wrapperv2.jsx
@@ -2,9 +2,6 @@ import React from "react";
 import Badge from "../../../components/Badge";
 import HeaderButtons from "../components/HeaderButtons";
 import Menu from "../components/Menu";
-import { canCreateMission } from "snu-lib";
-import { useSelector } from "react-redux";
-import { useHistory } from "react-router-dom";
 
 export default function Wrapper({ tab, structure, children, actionButton }) {
   if (!structure) return null;
@@ -24,23 +21,6 @@ export default function Wrapper({ tab, structure, children, actionButton }) {
         {tab === "details" && <HeaderButtons structure={structure} />}
       </div>
       <main className="mx-8 mt-6 mb-16">{children}</main>
-    </div>
-  );
-}
-
-export function ActionButton() {
-  const user = useSelector((state) => state.Auth.user);
-  const history = useHistory();
-
-  if (!canCreateMission(user)) {
-    return null;
-  }
-
-  return (
-    <div className="flex items-center justify-end">
-      <button className="cursor-pointer rounded-lg bg-blue-600 px-3 py-2 text-white" onClick={() => history.push(`/mission/create/${user.structureId}`)}>
-        Nouvelle mission
-      </button>
     </div>
   );
 }
